fix(Plugins): Import correctly files from directory when using adddir plugin
